Initial commit.
This commit contains a fork of the original InvoicePlane v2 software, which was originally released under MIT license by FusionInvoice and was supposed to be used as a basis for InvoicePlane v2.

As InvoicePlane v2 is still in the works, and I don't agree with the direction that (commercial) FusionInvoice is moving, I have decided to create this fork to support the 2 organisations that still use InvoicePlane v2.

This repository will likely only contain security updates.